--- pastlink-nodes/supabase/config.toml
@@ -1,1 +1,0 @@
-project_id = "uimmjzuqdqxfqoikcexf"
