+++ supabase/config.toml
@@ -1,4 +1,1 @@
-
-project_id = "kwzlxnfftzeqbtptyqzz"
-[functions.analyze-text]
-verify_jwt = false
+project_id = "uimmjzuqdqxfqoikcexf"
