+++ supabase/functions/analyze-node/index.ts
@@ -1,4 +1,3 @@
-
 import { GoogleGenerativeAI } from 'https://esm.sh/@google/generative-ai@^0.3.0'
 
 const corsHeaders = {
