+++ supabase/functions/analyze-text/tsconfig.json
