+++ supabase/functions/analyze-text/utils.ts
